+++ crates/rust-client/src/store/note_record/mod.rs
@@ -16,9 +16,9 @@
 //! [InputNoteRecord]'s and [OutputNoteRecord]'s documentation for more details about this.
 
 use alloc::string::{String, ToString};
-use core::fmt;
 
 use miden_objects::NoteError;
+use thiserror::Error;
 
 mod input_note_record;
 mod output_note_record;
@@ -36,40 +36,26 @@ pub mod input_note_states {
 // ================================================================================================
 
 /// Errors generated from note records.
-#[derive(Debug)]
+#[derive(Debug, Error)]
 pub enum NoteRecordError {
     /// Error generated during conversion of note record.
+    #[error("note record conversion error: {0}")]
     ConversionError(String),
     /// Invalid underlying note object.
-    NoteError(NoteError),
+    #[error("note error")]
+    NoteError(#[from] NoteError),
     /// Note record is not consumable.
+    #[error("note not consumable: {0}")]
     NoteNotConsumable(String),
     /// Invalid inclusion proof.
+    #[error("invalid inclusion proof")]
     InvalidInclusionProof,
     /// Invalid state transition.
+    #[error("invalid state transition: {0}")]
     InvalidStateTransition(String),
     /// Error generated during a state transition.
+    #[error("state transition error: {0}")]
     StateTransitionError(String),
-}
-
-impl fmt::Display for NoteRecordError {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        use NoteRecordError::*;
-        match self {
-            ConversionError(msg) => write!(f, "Note record conversion error: {}", msg),
-            NoteError(err) => write!(f, "Note error: {}", err),
-            NoteNotConsumable(msg) => write!(f, "Note not consumable: {}", msg),
-            InvalidInclusionProof => write!(f, "Invalid inclusion proof"),
-            InvalidStateTransition(msg) => write!(f, "Invalid state transition: {}", msg),
-            StateTransitionError(msg) => write!(f, "State transition error: {}", msg),
-        }
-    }
-}
-
-impl From<NoteError> for NoteRecordError {
-    fn from(error: NoteError) -> Self {
-        NoteRecordError::NoteError(error)
-    }
 }
 
 impl From<NoteRecordError> for String {